+++ eva/src/components/dash-filter-panel/FilterPart.vue
@@ -1,96 +1,75 @@
 <template>
-  <div class="pa-1 filter-part">
-    <v-row>
-      <v-col>
-        <div v-if="filterPart.filterPartType === 'token'">
-          <h4 :style="{ color: theme.$main_text }">
-            {{ elemName }}
-          </h4>
-          <div class="align-center d-flex">
-            <h5 :style="{ color: theme.$secondary_text }">
-              {{ filterPart.fieldName }}
-              ({{ filterPartValues.length }})
-            </h5>
-            <v-menu
-              offset-y
-              :close-on-content-click="false"
-              max-height="300"
-              v-if="filterPartValues.length > 0"
+  <div class="part-container">
+    <div class="data-col">
+      <template v-if="filterPart.filterPartType === 'token'">
+        <div class="part-title" v-text="elemName" />
+        <div class="part-subtitle">
+          <span>{{ filterPart.fieldName }} ({{ filterPartValues.length }})</span>
+          <v-menu
+            v-if="filterPartValues.length > 0"
+            offset-y
+            max-height="300"
+            :close-on-content-click="false"
+          >
+            <template v-slot:activator="{ on }">
+              <v-icon
+                :color="theme.$secondary_text"
+                right
+                size="18"
+                v-on="on"
+                v-text="dropDownIcon"
+              />
+            </template>
+            <v-list
+              class="part-values-list"
+              :style="{ color: theme.$main_text, backgroundColor: theme.$main_bg }"
             >
-              <template v-slot:activator="{ on, attrs }">
-                <v-btn
-                  icon
-                  x-small
-                  v-bind="attrs"
-                  v-on="on"
-                  :style="{ color: theme.$secondary_text }"
-                >
-                  <v-icon>{{ dropDownIcon }}</v-icon>
-                </v-btn>
-              </template>
-              <v-list
-                :style="{
-                  'background-color': theme.$main_bg,
-                  border: `1px solid ${theme.$secondary_border}`,
-                }"
-              >
-                <v-list-item
-                  v-for="(value, index) in filterPartValues"
-                  :key="index"
-                  class="align-center d-flex"
-                >
-                  <h5 :style="{ color: theme.$main_text }">
-                    {{ value }}
-                  </h5>
-                  <v-spacer></v-spacer>
-                  <v-btn icon x-small class="ml-2" v-if="isFocused" @click="removeValue(index)">
-                    <v-icon>{{ closeIcon }}</v-icon>
-                  </v-btn>
-                </v-list-item>
-              </v-list>
-            </v-menu>
-          </div>
+              <v-list-item v-for="(value, index) in filterPartValues" :key="index">
+                <span class="flex-grow-1" v-text="value"/>
+                <v-icon
+                  v-if="isFocused"
+                  :color="theme.$main_text"
+                  right
+                  size="20"
+                  @click="removeValue(index)"
+                  v-text="closeIcon"
+                />
+              </v-list-item>
+            </v-list>
+          </v-menu>
         </div>
+      </template>
+      <template v-else>
+        <div class="part-title">{{ filterPart.fieldName }}: {{ filterPart.value }}</div>
+        <div class="part-subtitle" v-text="`(${filterPart.fieldType}) ${operationManualTitle}`" />
+      </template>
+    </div>
 
-        <div v-else>
-          <h4 :style="{ color: theme.$main_text }">
-            {{ filterPart.fieldName }}: {{ filterPart.value }}
-          </h4>
-          <h5 :style="{ color: theme.$secondary_text }">
-            ({{ filterPart.fieldType }}) {{ operationManualTitle }}
-          </h5>
-        </div>
-      </v-col>
-
-      <v-col class="d-flex flex-column align-center justify-end" v-if="isFocused">
-        <v-btn
-          icon
-          x-small
-          v-show="filterPart.filterPartType === 'token'"
-          :color="theme.$main_text"
-          @click.stop.prevent="$store.commit('refreshFilterPart', { idDash, filterIndex, filterPartIndex })"
-        >
-          <v-icon>{{ refreshIcon }}</v-icon>
-        </v-btn>
-        <v-btn
-          icon
-          x-small
-          v-if="editPermission || filterPart.filterPartType === 'manual'"
-          class="justify-end"
-          :color="theme.$error_color"
-          @click.stop.prevent="
-            $store.commit('deleteFilterPart', { idDash, filterIndex, filterPartIndex })
-          "
-        >
-          <v-icon>{{ closeIcon }}</v-icon>
-        </v-btn>
-      </v-col>
-    </v-row>
+    <div v-if="isFocused" class="button-col">
+      <v-icon
+        v-if="filterPart.filterPartType === 'token'"
+        :color="theme.$main_text"
+        size="20"
+        @click.stop.prevent="
+          $store.commit('refreshFilterPart', { idDash, filterIndex, filterPartIndex })
+        "
+        v-text="refreshIcon"
+      />
+      <v-icon
+        v-if="editPermission || filterPart.filterPartType === 'manual'"
+        :color="theme.$error_color"
+        size="20"
+        @click.stop.prevent="
+          $store.commit('deleteFilterPart', { idDash, filterIndex, filterPartIndex })
+        "
+        v-text="closeIcon"
+      />
+    </div>
   </div>
 </template>
 
 <script>
-  import { mdiClose, mdiRefresh, mdiPencil, mdiChevronDown } from '@mdi/js';
+  import { mdiCloseCircleOutline, mdiRefresh, mdiPencil, mdiChevronDown } from '@mdi/js';
 
   export default {
     name: 'FilterPart',
@@ -104,7 +83,7 @@
     ],
     data() {
       return {
-        closeIcon: mdiClose,
+        closeIcon: mdiCloseCircleOutline,
         refreshIcon: mdiRefresh,
         editIcon: mdiPencil,
         dropDownIcon: mdiChevronDown,
@@ -156,9 +135,36 @@
   };
 </script>
 
-<style scoped>
-  .filter-part {
-    border-right: 1px solid rgba(0, 0, 0, 0.12);
-    border-radius: 0;
-  }
+<style lang="sass" scoped>
+.part-container
+  display: flex
+  height: 100%
+
+  .data-col, .button-col
+    display: flex
+    flex-direction: column
+    justify-content: space-evenly
+
+  .data-col
+    min-width: 100px
+
+    .part-title
+      color: var(--main_text)
+      font-weight: 600
+      white-space: nowrap
+      overflow: hidden
+      text-overflow: ellipsis
+
+    .part-subtitle
+      display: flex
+      align-items: center
+      color: var(--secondary_text)
+      font-size: 14px
+      white-space: nowrap
+      overflow: hidden
+      text-overflow: ellipsis
+
+  .button-col
+    align-items: center
+    margin-left: 20px
 </style>